# book-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "bill",
      "url": "https://example.com",
      "email": "bill@example.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://example-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local"
    },
    {
      "url": "https://production.com",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/auth/authenticate": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "authentication",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "post": {
        "tags": [
          "Feedback"
        ],
        "operationId": "saveFeedback_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/book/{book-id}": {
      "get": {
        "tags": [
          "Feedback"
        ],
        "operationId": "findAllFeedbackBook",
        "parameters": [
          {
            "name": "book-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFeedbackResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthenticationRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "maxLength": 2147483647,
            "minLength": 8,
            "type": "string"
          }
        }
      },
      "AuthenticationResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "FeedbackRequest": {
        "required": [
          "bookId",
          "comment"
        ],
        "type": "object",
        "properties": {
          "note": {
            "maximum": 5,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "comment": {
            "type": "string"
          },
          "bookId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "FeedbackResponse": {
        "type": "object",
        "properties": {
          "note": {
            "type": "number",
            "format": "double"
          },
          "comment": {
            "type": "string"
          },
          "ownFeedback": {
            "type": "boolean"
          }
        }
      },
      "PageResponseFeedbackResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedbackResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElement": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}